requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
commonmark==0.9.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.3
openai==0.16.0
openpyxl==3.0.9
pandas==1.4.1
pandas-stubs==1.2.0.53
pyfiglet==0.8.post1
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
termcolor==1.1.0
tqdm==4.63.1
urllib3==1.26.9
